# gh-118761: Optimise import time for ``string``