This PR achieves a 27x improvement in import time for the `string` module. The main improvement comes from replacing `Template.__init_subclass__()` (GH-16256) with a descriptor class, allowing lazy import of the `re` module.

Current:

```
import string: cumulative time
mean: 9162.100 µs
median: 9133.000 µs
stdev: 66.662
min: 9071
max: 9301
```

This PR:

```
import string: cumulative time
mean: 334.967 µs
median: 329.000 µs
stdev: 13.438
min: 316
max: 368
```

<!-- gh-issue-number: gh-118761 -->
* Issue: gh-118761
<!-- /gh-issue-number -->